feat: 易支付(EPay) v1/v2 协议兼容层

+++ app/controller/Index.php
@@ -487,22 +487,32 @@ class Index extends BaseController
                 return json($this->getReturn(-1, "订单已过期"));
             }
 
-            $key = Setting::getConfigValue("key");
+            $orderData = $res->toArray();
 
-            $res['price'] = number_format((float)$res['price'], 2, ".", "");
-            $res['really_price'] = number_format((float)$res['really_price'], 2, ".", "");
-
-            $p = "payId=" . $res['pay_id'] . "&param=" . $res['param'] . "&type=" . $res['type'] . "&price=" . $res['price'] . "&reallyPrice=" . $res['really_price'];
-
-            $sign = $res['pay_id'] . $res['param'] . $res['type'] . $res['price'] . $res['really_price'] . $key;
-            $p = $p . "&sign=" . md5($sign);
-
-            $url = $res['return_url'];
-
-            if (strpos($url, "?") === false) {
-                $url = $url . "?" . $p;
+            if (\app\service\epay\EpayNotifyService::isEpayOrder($orderData)) {
+                $epayConfig = \app\service\epay\EpayConfigService::getConfig();
+                $signingKey = \app\service\epay\EpayNotifyService::isEpayV2Order($orderData)
+                    ? $epayConfig['private_key']
+                    : $epayConfig['key'];
+                $url = \app\service\epay\EpayNotifyService::buildReturnUrl($orderData, $signingKey);
             } else {
-                $url = $url . "&" . $p;
+                $key = Setting::getConfigValue("key");
+
+                $res['price'] = number_format((float)$res['price'], 2, ".", "");
+                $res['really_price'] = number_format((float)$res['really_price'], 2, ".", "");
+
+                $p = "payId=" . $res['pay_id'] . "&param=" . $res['param'] . "&type=" . $res['type'] . "&price=" . $res['price'] . "&reallyPrice=" . $res['really_price'];
+
+                $sign = $res['pay_id'] . $res['param'] . $res['type'] . $res['price'] . $res['really_price'] . $key;
+                $p = $p . "&sign=" . md5($sign);
+
+                $url = $res['return_url'];
+
+                if (strpos($url, "?") === false) {
+                    $url = $url . "?" . $p;
+                } else {
+                    $url = $url . "&" . $p;
+                }
             }
 
             return json($this->getReturn(1, "成功", $url));
@@ -605,27 +615,43 @@ class Index extends BaseController
             ->find();
 
         if ($res) {
-            TmpPrice::where("oid", $res['order_id'])->delete();
+            $affected = PayOrder::where("id", $res['id'])->where("state", 0)
+                ->update(array("state" => 1, "pay_date" => time(), "close_date" => time()));
 
-            PayOrder::where("id", $res['id'])->update(array("state" => 1, "pay_date" => time(), "close_date" => time()));
-
-            $url = $res['notify_url'];
-
-            $key = Setting::getConfigValue("key");
-
-            $p = "payId=" . $res['pay_id'] . "&param=" . $res['param'] . "&type=" . $res['type'] . "&price=" . $res['price'] . "&reallyPrice=" . $res['really_price'];
-
-            $sign = $res['pay_id'] . $res['param'] . $res['type'] . $res['price'] . $res['really_price'] . $key;
-            $p = $p . "&sign=" . md5($sign);
-
-            if (strpos($url, "?") === false) {
-                $url = $url . "?" . $p;
-            } else {
-                $url = $url . "&" . $p;
+            if ($affected === 0) {
+                return json($this->getReturn(1, "订单已处理"));
             }
 
-            $re = $this->getCurl($url);
-            if ($re == "success") {
+            TmpPrice::where("oid", $res['order_id'])->delete();
+
+            $orderData = $res->toArray();
+
+            if (\app\service\epay\EpayNotifyService::isEpayOrder($orderData)) {
+                $epayConfig = \app\service\epay\EpayConfigService::getConfig();
+                $signingKey = \app\service\epay\EpayNotifyService::isEpayV2Order($orderData)
+                    ? $epayConfig['private_key']
+                    : $epayConfig['key'];
+                $notifyOk = \app\service\epay\EpayNotifyService::sendNotify($orderData, $signingKey);
+            } else {
+                $url = $res['notify_url'];
+
+                $key = Setting::getConfigValue("key");
+
+                $p = "payId=" . $res['pay_id'] . "&param=" . $res['param'] . "&type=" . $res['type'] . "&price=" . $res['price'] . "&reallyPrice=" . $res['really_price'];
+
+                $sign = $res['pay_id'] . $res['param'] . $res['type'] . $res['price'] . $res['really_price'] . $key;
+                $p = $p . "&sign=" . md5($sign);
+
+                if (strpos($url, "?") === false) {
+                    $url = $url . "?" . $p;
+                } else {
+                    $url = $url . "&" . $p;
+                }
+
+                $notifyOk = $this->getCurl($url) == "success";
+            }
+
+            if ($notifyOk) {
                 return json($this->getReturn());
             } else {
                 PayOrder::where("id", $res['id'])->update(array("state" => 2));
